[Partners] Extract DraftJS editor into the separate RichTextEditor component
